Split the camera into header and cpp file
