+++ include/road_network/SkyMaps/stb_image.h
