+++ include/road_network/output_main.hpp
@@ -17,7 +17,7 @@
 
 #include "road_network/output_gl_framework.hpp"
 #include "road_network/shader_util.hpp"
-#include "road_network/camera.h"
+#include "road_network/Camera/camera.hpp"
 #include "road_network/Road-Model/road_collection.hpp"
 #include "road_network/Traffic-Manager/manager.hpp"
 #include "road_network/SkyMaps/skymaps.hpp"
